feat(gui): add input field for username
